move cli to examples folder.
