+++ client/examples/cli.yml
