--- products/commandline/.license_template
@@ -1,2 +1,0 @@
-// Copyright {20\d{2}(-20\d{2})?} IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
