--- products/commandline/gon-config.json
@@ -1,13 +1,0 @@
-{
-  "source": ["./target/release/stronghold"],
-  "bundle_id": "org.iota.stronghold-cli",
-  "apple_id": {
-    "password": "@env:AC_PASSWORD"
-  },
-  "sign": {
-    "application_identity": "Developer ID Application: IOTA Stiftung (UG77RJKZHH)"
-  },
-  "zip": {
-    "output_path": "./stronghold.zip"
-  }
-}
